move awx.main.expect to awx.main.isolated
