+++ awx/main/isolated/manager.py
